backend/requirements.txt
# TinyDictate Backend Requirements
# Python 3.11+ recommended

# Core dependencies
torch>=2.1.0
numpy>=1.24.0

# NeMo ASR toolkit
nemo_toolkit[asr]>=1.22.0

# WebSocket server
websockets>=12.0

# Hugging Face model hub
huggingface_hub>=0.20.0

# Speaker diarization
pyannote.audio>=3.1.0

# Optional: For better audio processing
# soundfile>=0.12.0
# librosa>=0.10.0
